+++ itachallenge-user/src/main/java/com/itachallenge/user/service/IChallengeService.java
@@ -1,0 +1,9 @@
+package com.itachallenge.user.service;
+
+import reactor.core.publisher.Mono;
+
+public interface IChallengeService {
+
+    Mono<Boolean> addChallengeToSolved(String challengeId);
+
+}